+++ src/algorithms/resampler/adapters/direct_resampler_conditioner.cc
@@ -51,7 +51,7 @@ DirectResamplerConditioner::DirectResamplerConditioner(
     std::string default_dump_file = "./data/signal_conditioner.dat";
     double fs_in;
     fs_in = configuration->property("GNSS-SDR.internal_fs_hz", 2048000.0);
-    sample_freq_in_ = configuration->property(role_ + ".sample_freq_in", (double)4000000.0);
+    sample_freq_in_ = configuration->property(role_ + ".sample_freq_in", 4000000.0);
     sample_freq_out_ = configuration->property(role_ + ".sample_freq_out", fs_in);
     if(std::fabs(fs_in - sample_freq_out_) > std::numeric_limits<double>::epsilon())
         {
